refactor(react-gui): move the common pages onto the schema
RendererCommonSection.tsx was two things at once: the page every renderer
shows first, and the library of row components that 15 other files import.
Describing the page as data means separating them.

The rows become a catalog of their own under rows/, one file per kind,
alongside the two that already lived there. The bodies move verbatim -- only
the doc headers change, since "exported so the type-specific sections can
reuse it" stops being the reason a row is public once it lives in a catalog.
NumEnumRow goes with the move: it has no consumers.

The three common pages then become schemas. They stay three, because the
nodes the inspector opens genuinely differ: a renderer has opacity, material
and edge lines; an object has a read-only source path instead; a renderer
group inherits the whole Renderer property set in C++ but draws nothing
(RendGroup::display is empty), so all it can honestly show is name, visible
and locked.

Two additions to the engine, both needed by the edge-lines block:

  - an `enum` kind, for a dropdown whose options read as their raw C++ ids.
    (nucl and isosurf want it too; `mappedEnum` with empty labels would have
    rendered the same but said the wrong thing.)
  - a `typeIs` predicate. Edge lines are suppressed for the line-only
    renderers, and those inherit egtype / eglinew / egcolor from the C++
    Renderer base regardless, so the gate has to be on the renderer type
    rather than on which properties exist.

PropertiesTab is now the composition and nothing else: pick the common
sections for this kind of node, append the type-specific ones, render.

The parity snapshots do not move -- which is the point, since the common page
is on every one of the 38 fixtures.
