+++ tritium/react-gui/src/renderer/components/inspector/rows/index.ts
@@ -1,0 +1,29 @@
+/**
+ * @file components/inspector/rows/index.ts
+ * @description The Properties tab's row catalog.
+ *
+ * One component per kind of property control. The schema engine
+ * (`SchemaSection`) picks one per row of a page's schema, and the pages not
+ * yet described as data compose them by hand; both go through this barrel so
+ * a row is defined once.
+ */
+
+export { resetProps } from './rowProps';
+export type { RowProps, SetFn, ResetFn } from './rowProps';
+export { TextRow } from './TextRow';
+export { BoolRow } from './BoolRow';
+export { NumRow } from './NumRow';
+export { NumInputRow } from './NumInputRow';
+export { SliderRow } from './SliderRow';
+export { EnumRow } from './EnumRow';
+export type { EnumRowProps } from './EnumRow';
+export { MappedEnumRow } from './MappedEnumRow';
+export type { MappedEnumRowProps } from './MappedEnumRow';
+export { ColorRow } from './ColorRow';
+export type { ColorRowProps } from './ColorRow';
+export { SelRow } from './SelRow';
+export type { SelRowProps } from './SelRow';
+export { AsyncSelectRow, useAsyncNames } from './AsyncSelectRow';
+export type { AsyncSelectRowProps, AsyncNameSource } from './AsyncSelectRow';
+export { OptionalNumRow } from './OptionalNumRow';
+export type { OptionalNumRowProps } from './OptionalNumRow';
